Add re2 dependency to libaddressinput.
This patch adds the re2 dependency to libaddressinput. The dependency will
be used to validate postal code format of addresses. The addresses are
provided by the user. The format regular expressions are from:

 https://i18napis.appspot.com/ssl-address

For example, the US postal code has the following regular expression:

 \\d{5}([ \\-]\\d{4})?

BUG=327046

Review URL: https://codereview.chromium.org/98543010