Add address metadata query functionality to libaddressinput.
Specifically the utility functions IsFieldRequired() and IsFieldUsed()
that answers the questions of whether a particular field is a required
field or a field that is used, respectively, for a particular region.

This is implemented through the class AddressValidator, which provides
the public interface, and the class MetadataQueryTask which encapsulates
the information necessary to perform the checking for an Is*() method
that answers a yes/no question and call a callback when that has been
done.

The interface is asynchronous, allowing it to be used with a Downloader
implementation that waits for several pending requests simultaneously.

By virtue of being part of the class AddressValidator, this shares the
metadata cache with address validation. (The future might show this to
be a not so good design decision, but with the actual implementation
isolated in class MetadataQueryTask it should be straightforward to move
things around then.)