Determine language tag and order of UI components.
This patch changes BuildComponents() function to read the UI language
tag and determine the language tag that should be stored with the
address data. The function also chooses between the default and the
Latin order of UI components in the address form.

A language tag is a sequence of one of more subtags separated by
hyphen ("-"). All characters are in ASCII and case-insensitive. The
first subtag is the base language, for example "zh". The first
four-character subtag is the script, for example "Hans".