Break out abstract base class Supplier from MetadataLoader.
To make it possible to supply the AddressValidator with address metadata
in several different way, an interface for such a Supplier is defined
and the existing MetadataLoader is refactored to implement it.

This will be used to implement another Supplier that doesn't load any
data on demand, but only works with preloaded data.

R=rouslan@chromium.org
BUG=

Review URL: https://codereview.appspot.com/97650044