Make the Callback templates more generic.
Instead of unconditionally adding const& to the specified types, require
the user of the Callback templates to specify it. This makes it possible
to use these same templates also for other types.

Now, const int& is replaced with just plain int. Later, this will be
used to replace const std::string& with std::string*.

R=rouslan@chromium.org