More natural language friendly string matching in PreloadSupplier.
This is done by updating the StringCompare helper class to also provide
a comparison method that conforms to the STL requirements for less<>
predicates, and then wrapping this as an STL predicate.

(A portable default implementation of this is done through "clever" use
of RE2 to get UTF-8 capable string comparison. Users of the library can
override this with better/faster implementations.)

This will make PreloadSupplier able to match "ZÜRICH" and "Zürich". Yay!
;-)

R=rouslan@chromium.org
BUG=

Review URL: https://codereview.appspot.com/101340044