Prepare for RapidJSON FindMember() API change.
Currently, ConstMemberIterator is implemented as a pointer and
FindMember() returns NULL to signal not found. But ConstMemberIterator
is going to be changed to a non-pointer type and the return value for
not found will then be changed to MemberEnd().

To make this code work for both cases, the return value is first
compared to a newly created empty ConstMemberIterator (ie. NULL, when
ConstMemberIterator is implemented as a pointer) and then compared to
MemberEnd().