Avoid manual lifetime management of `std::string`
This change touches several places in `Source` and `Storage` where
string was being passed as a pointer, with expectations for `delete` to
be called once done with the value. There may even be a case in Chromium
at the moment where one of these calls may be leaking memory due to how
the memory is expected to managed.

This change passes the string by value, and in cases where the value may
be omitted the string is wrapped on a `std::optional`.

Bug: 423542167